// src/types/delegation.ts
export type IDelegationItem = {
  id: number;
  delegationScope: string;
  fromUser: number;
  toUser: number;
  dateFrom: string;
  dateTo: string;
  toUserName: string;
  fromUserName: string;
  avatarUrl?: string;
  active: number;
  activeName: string;
  createdDate: string;
  updatedDate: string;
  createdByUserName: string;
  updatedByUserName: string;
  delegationScopeName: string;
  // id: string;
  // fromUser: string;
  // fromUserid: number;
  // toUser: string;
  // toUserid: number;
  // dateFrom: Date;
  // dateTo: Date;
  // scope: number;
  // createdAt: Date;
  // createdBy: string;
  // updatedAt: Date;
  // updatedBy: string;
  // process?: string[] | null;
};
export type IDelegationInfo = {
  id?: number;
  delegationScopeDTO?: any[];
  active: number;
  fromUser: string;
  toUser: string;
  delegationScope: string;
  dateFrom: Date;
  dateTo: Date;
  toUserName?: string;
  fromUserName?: string;
};
export type IDelegationInfomap = {
  id?: number;
  delegationScopeDTO?: any[];
  active: number;
  fromUser: string;
  toUser: string;
  delegationScope: number;
  dateFrom: string | Date|null;
  dateTo: string | Date|null;
  toUserName?: string;
  fromUserName?: string;
};
export type IDelegationTableFilterValue = string | string[];
export type IDelegationTableFilters = {
  fromUserName: string;
  activeName: string;
};